metallics_optimizer/requirements.txt
Django==3.1.2
djangorestframework==3.12.1
markdown==3.3.2
django-filter==2.4.0
